Fix CoM crash when deleting cipher.
When deleting a cipher, the cipher doesn't disappear – trying to
delete it again causes the application to crash.  This fix makes the
list behind ciphers mutable, and adds/removes as needed.  It also
fixes the various labels to indicate “Cypher” rather than notes.

This issue probably also shows in other titles, I haven't checked yet,
but the various Note-like custom fields should probably be made
generic to handle this in a nicer way.